# stuplus.api/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Stuplus API",
    "description": "Stuplus API Documentation."
  },
  "servers": [
    {
      "url": "http://212.98.224.208:25010/",
      "description": "Canli API"
    },
    {
      "url": "http://127.0.0.1:25010/",
      "description": "Local API"
    }
  ],
...
  "paths": {
...
    "/login/": {
      "post": {
        "tags": [
          "Login"
        ],
        "description": "Login.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        }
      }
    },
    "/login/google": {
      "post": {
        "tags": [
          "Login"
        ],
        "description": "Login with google(also registers if user not exist).",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginGoogleResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/LoginGoogleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginGoogleRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/LoginGoogleRequest"
              }
            }
          }
        }
      }
    },
    "/login/register": {
      "post": {
        "tags": [
          "Login"
        ],
        "description": "Register.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginRegisterResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/LoginRegisterResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRegisterRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/LoginRegisterRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "stuplus@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "tesTtesT@33"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "xml": {
          "name": "LoginRequest"
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2MmFiOGEyMDQxNjZmZDFlYWViYmIzZmEiLCJyb2xlIjowLCJpYXQiOjE2NTYwNjA0MzAsImV4cCI6MTY1ODY1MjQzMH0.ZwB2RiwHwRx_-HkU1ziIwWeXQucaPQFBDQH5QVjjvdE"
          }
        },
        "xml": {
          "name": "LoginResponse"
        }
      },
      "LoginGoogleRequest": {
        "type": "object",
        "properties": {
          "AccessToken": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "AccessToken"
        ],
        "xml": {
          "name": "LoginGoogleRequest"
        }
      },
      "LoginGoogleResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2MmFiOGEyMDQxNjZmZDFlYWViYmIzZmEiLCJyb2xlIjowLCJpYXQiOjE2NTYwNjA0MzAsImV4cCI6MTY1ODY1MjQzMH0.ZwB2RiwHwRx_-HkU1ziIwWeXQucaPQFBDQH5QVjjvdE"
          }
        },
        "xml": {
          "name": "LoginGoogleResponse"
        }
      },
      "LoginRegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "stuplus@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "tesTtesT@33"
          },
          "gender": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "enum"
              },
              "values": {
                "type": "object",
                "properties": {
                  "NotSpecified": {
                    "type": "number",
                    "example": 0
                  },
                  "Male": {
                    "type": "number",
                    "example": 1
                  },
                  "Female": {
                    "type": "number",
                    "example": 2
                  },
                  "NotDefined": {
                    "type": "number",
                    "example": 3
                  }
                }
              }
            }
          }
        },
        "required": [
          "email",
          "password",
          "gender"
        ],
        "xml": {
          "name": "LoginRegisterRequest"
        }
      },
      "LoginRegisterResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2MmFiOGEyMDQxNjZmZDFlYWViYmIzZmEiLCJyb2xlIjowLCJpYXQiOjE2NTYwNjA0MzAsImV4cCI6MTY1ODY1MjQzMH0.ZwB2RiwHwRx_-HkU1ziIwWeXQucaPQFBDQH5QVjjvdE"
          }
        },
        "xml": {
          "name": "LoginRegisterResponse"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}